requirements.txt
flask==2.0.1
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.5.3.56
numpy==1.21.2
pillow==8.3.2
werkzeug==2.0.1
tensorboardX==2.4
scipy==1.7.1
tqdm==4.62.3 